fix: align qwen3 4b 2507 with genie

+++ core/include/llm/llm_utils.h
@@ -20,13 +20,13 @@ class GENIEX_API RotaryEmbedding {
     RotaryEmbedding(size_t head_dim, float theta = 10000.f);
 
     // Returns {cos, sin}, each flat [n * half_dim] where n = position_ids.size().
-    std::pair<std::vector<float>, std::vector<float>> forward(const std::vector<int32_t>& position_ids) const;
+    std::pair<std::vector<double>, std::vector<double>> forward(const std::vector<int32_t>& position_ids) const;
 
     size_t halfDim() const;
 
    private:
-    std::vector<float> inv_freq_;  // [half_dim]
-    size_t             half_dim_ = 0;
+    std::vector<double> inv_freq_;  // [half_dim]
+    size_t              half_dim_ = 0;
 };
 
 // LongRoPE with dynamic scaling and per-dimension extension factors.
@@ -37,17 +37,17 @@ class GENIEX_API LongRoPEEmbedding {
     LongRoPEEmbedding(size_t head_dim, float theta, std::vector<float> ext_factors,
         int max_position_embeddings = 131072, int original_max_position_embeddings = 4096);
 
-    std::pair<std::vector<float>, std::vector<float>> forward(const std::vector<int32_t>& position_ids) const;
+    std::pair<std::vector<double>, std::vector<double>> forward(const std::vector<int32_t>& position_ids) const;
 
     size_t halfDim() const;
 
    private:
-    std::vector<float> ext_factors_;  // [half_dim], per-dimension extension factors
-    float              base_                             = 10000.f;
-    int                dim_                              = 0;  // full head_dim (half_dim_ * 2)
-    int                max_position_embeddings_          = 131072;
-    int                original_max_position_embeddings_ = 4096;
-    size_t             half_dim_                         = 0;
+    std::vector<double> ext_factors_;  // [half_dim], per-dimension extension factors
+    double              base_                             = 10000.0;
+    int                 dim_                              = 0;  // full head_dim (half_dim_ * 2)
+    int                 max_position_embeddings_          = 131072;
+    int                 original_max_position_embeddings_ = 4096;
+    size_t              half_dim_                         = 0;
 };
 
 // Partial RoPE: rotates only (rope_fraction * head_dim) dimensions, with a post-scale factor.
@@ -56,21 +56,21 @@ class GENIEX_API PartialRoPEEmbedding {
     PartialRoPEEmbedding() = default;
     PartialRoPEEmbedding(size_t head_dim, float theta = 10000.f, float rope_fraction = 1.0f, float scale = 1.0f);
 
-    std::pair<std::vector<float>, std::vector<float>> forward(const std::vector<int32_t>& position_ids) const;
+    std::pair<std::vector<double>, std::vector<double>> forward(const std::vector<int32_t>& position_ids) const;
 
     size_t halfDim() const;
 
    private:
-    std::vector<float> inv_freq_;  // [rope_half_dim]
-    float              scale_         = 1.f;
-    size_t             rope_half_dim_ = 0;
+    std::vector<double> inv_freq_;  // [rope_half_dim]
+    double              scale_         = 1.0;
+    size_t              rope_half_dim_ = 0;
 };
 
 // Returns position IDs [n_past, n_past + count) as a flat int32 vector.
 GENIEX_API std::vector<int32_t> get_position_ids(size_t n_past, size_t count);
 
 // Returns {cos, sin}, each flat [n * half_dim] for the given position IDs.
-GENIEX_API std::pair<std::vector<float>, std::vector<float>> get_cos_sin(
+GENIEX_API std::pair<std::vector<double>, std::vector<double>> get_cos_sin(
     const std::vector<int32_t>& position_ids, size_t head_dim, float rope_theta = 10000.f);
 
 // Returns a causal attention mask, flat [seq_len * (kv_len + seq_len)].
